+++ src/actions/message.js
@@ -29,5 +29,5 @@ export const list = createAction(LIST, (msgs) => msgs);
 export const OLD = 'MESSAGE_OLD';
 export const old = createAction(OLD, (msgs) => msgs);
 
-export const IMAGE = 'MESSAGE_IMAGE';
-export const image = createAction(IMAGE, (msg) => msg, sync);
+export const FILE = 'MESSAGE_FILE';
+export const file = createAction(FILE, (msg) => msg, sync);