+++ src/containers/MessageForm.js
@@ -1,7 +1,7 @@
 /* eslint camelcase: "off" */
 
 import { open } from '../actions/dialog';
-import { create as createMessage, image } from '../actions/message';
+import * as Message from '../actions/message';
 import { update } from '../actions/typing';
 import { create, remove } from '../actions/name';
 import { connect } from 'react-redux';
@@ -17,9 +17,9 @@ export default connect(
         onCreateName: (e, name) => dispatch(create(name)),
         onEditName: (e, name_id) => dispatch(open('name-edit', { name_id })),
         onRemoveName: (e, id) => dispatch(remove({ id })),
-        onSendMessage: (e, message) => dispatch(createMessage(message)),
+        onSendMessage: (e, message) => dispatch(Message.create(message)),
         onTyping: (e, name, message) => dispatch(update({ name, message })),
-        onUploadImage: (e, name, file) => dispatch(image({
+        onUploadImage: (e, name, file) => dispatch(Message.file({
             character_url: name.character_url,
             icon_id: name.icon_id,
             name: name.name,