+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - d6 animation for Fluorite5 script
 - Dice results tooltip for Fluorite5 script
 - Fluorite5 mods
+- Audio file uploading
 
 ### Fixed
 - Improve performance of name edit dialog